Initial implementation of in-memory Lru Cache on lib-cache to reduce the number of database transactions/queries that will occur. Paves the way for implementing moving response body data based on whether or not it has a lock.

+++ core/src/commonMain/kotlin/com/ustadmobile/core/domain/blob/savelocaluris/SaveLocalUrisAsBlobsUseCase.kt
@@ -50,7 +50,7 @@ interface SaveLocalUrisAsBlobsUseCase {
         val entityUid: Long,
         val localUri: String,
         val blobUrl: String,
-        val retentionLockId: Int = 0,
+        val retentionLockId: Long = 0,
         val integrity: String,
         val mimeType: String,
         val storageSize: Long,
