+++ lib-database/src/commonMain/kotlin/com/ustadmobile/lib/db/entities/CacheLockJoin.kt
@@ -19,7 +19,7 @@ data class CacheLockJoin(
     var cljEntityUid: Long = 0,
     //This should not be nullable, but door doesn't support that at the moment. In reality it will never be null
     var cljUrl: String? = "",
-    var cljLockId: Int = 0,
+    var cljLockId: Long = 0,
     var cljStatus: Int = 0,
     var cljType: Int = 0,
 ) {
